requirements.txt
pandas
numpy<2
scikit-learn
faiss-cpu==1.8.0
sentence-transformers==2.7.0
tqdm
fastapi
uvicorn